requirements.txt
Appium-Python-Client==2.8.1
openpyxl==3.1.0
confparser==0.1.7
pytest==7.2.1
allure-pytest==2.12.0
PyHamcrest==2.0.4
randomtimestamp==2.2

